0.2.0 Update
Added a property for automodule called :hide-output:. When it is set to
True, it will generate no output for the module in the html. However,
the module is still processed, so the links to it work and when used the
description is printed. Useful for modules that only contain other
submodules, the routines and variables of which they use.