0.2.1 Update#
Fixed a freak-occurrence bug. The way the variable lines are parsed is,
the code looks for type,dimension,variable,comment. However the regex
is a bit of a mess, and every line in the code is parsed. If a code line
has a comment at the end, and  before that there's a variable name, that
may be overwrite the already parsed variable/comment combination. Since
the variables are declared first thing in a function, fixed by only
taking the first line that matches the regex for a given variable name.