0.5.4 Update
Properly account for bindings in variables and functions.
Backup description if not accessible.
Add the possibility of describing a variable in the function description
using the keyword "@param $name:" at the beginning of lines. Those lines
will be removed from the comment. All lines with the same $name will be
added to the description of the parameter $name